+++ src/components/views/analytics/MarketMetrics.tsx
@@ -102,7 +102,7 @@ export const MarketMetrics = ({ marketMetricsData }: MarketMetricsProps) => {
   const { theme } = useTheme();
 
   const chartColors = useChartColors(
-    theme as "charcoal" | "midnight" | "obsidian" | "snowlight"
+    theme as "dark" | "light"
   );
 
   return (