+++ src/components/views/analytics/YearOverview.tsx
@@ -126,7 +126,7 @@ export const YearOverview = ({ yearOverviewData }: YearOverviewProps) => {
   const { theme } = useTheme();
 
   const chartColors = useChartColors(
-    theme as "charcoal" | "midnight" | "obsidian" | "snowlight"
+    theme as "dark" | "light"
   );
 
   const { width: windowWidth } = useWindowDimensions();