+++ src/components/views/homepage/BestSellingProducts.tsx
@@ -99,7 +99,7 @@ export const BestSellingProducts = ({
   const { theme } = useTheme();
 
   const chartColors = useChartColors(
-    theme as "charcoal" | "midnight" | "obsidian" | "snowlight"
+    theme as "dark" | "light"
   );
 
   const { width: windowWidth } = useWindowDimensions();