+++ src/hooks/useChartColors.ts
@@ -14,38 +14,10 @@ interface ChartColors {
 }
 
 export const useChartColors = (
-  theme: "charcoal" | "midnight" | "obsidian" | "snowlight"
+  theme: "dark" | "light"
 ): ChartColors => {
   const themeColors = {
-    charcoal: {
-      primary: {
-        stroke: "rgb(61, 185, 133)",
-        fill: "rgb(98, 136, 227)",
-        disabled: "rgb(255,255,255,0.1)",
-        grid: "rgba(255,255,255,0.1)",
-      },
-      secondary: {
-        stroke: "hsl(199.4, 13.3%, 45.7%)",
-        fill: "hsl(199.4, 13.3%, 45.7%)",
-        disabled: "rgb(255,255,255,0.1)",
-        grid: "rgba(255,255,255,0.1)",
-      },
-    },
-    midnight: {
-      primary: {
-        stroke: "rgb(6, 182, 212)",
-        fill: "rgb(6, 182, 212)",
-        disabled: "rgb(255,255,255,0.1)",
-        grid: "rgba(255,255,255,0.1)",
-      },
-      secondary: {
-        stroke: "rgb(6, 182, 212)",
-        fill: "rgb(255,255,255,0.2)",
-        disabled: "rgb(255,255,255,0.1)",
-        grid: "rgba(255,255,255,0.1)",
-      },
-    },
-    obsidian: {
+    dark: {
       primary: {
         stroke: "rgb(61, 185, 133)",
         fill: "rgb(61, 185, 133)",
@@ -59,7 +31,7 @@ export const useChartColors = (
         grid: "rgba(255,255,255,0.1)",
       },
     },
-    snowlight: {
+    light: {
       primary: {
         stroke: "rgb(118, 167, 247)",
         fill: "rgb(118, 167, 247)",
@@ -77,19 +49,15 @@ export const useChartColors = (
 
   if (!theme) {
     if (typeof document !== "undefined") {
-      if (document.documentElement.classList.contains("obsidian")) {
-        return themeColors.obsidian;
-      } else if (document.documentElement.classList.contains("snowlight")) {
-        return themeColors.snowlight;
-      } else if (document.documentElement.classList.contains("midnight")) {
-        return themeColors.midnight;
-      } else if (document.documentElement.classList.contains("charcoal")) {
-        return themeColors.charcoal;
+      if (document.documentElement.classList.contains("dark")) {
+        return themeColors.dark;
+      } else if (document.documentElement.classList.contains("light")) {
+        return themeColors.light;
       }
     }
 
-    return themeColors.obsidian;
+    return themeColors.dark;
   }
 
-  return themeColors[theme] || themeColors.obsidian;
+  return themeColors[theme] || themeColors.dark;
 };