+++ src/services/providers.tsx
@@ -7,7 +7,7 @@ import { ClerkProvider } from "@clerk/nextjs";
 import { Layout } from "../layout/Layout";
 import { client } from "./apolloClient";
 
-export const THEMES_ARRAY = ["snowlight", "midnight", "charcoal", "obsidian"];
+export const THEMES_ARRAY = ["light", "dark"];
 
 export const Providers = ({ children }: { children: React.ReactNode }) => {
   return (
@@ -19,7 +19,7 @@ export const Providers = ({ children }: { children: React.ReactNode }) => {
           enableSystem={false}
           attribute="class"
           themes={THEMES_ARRAY}
-          defaultTheme="obsidian"
+          defaultTheme="dark"
           disableTransitionOnChange
         >
           <Layout>{children}</Layout>